+++ src/stringUtils.js
